current_pip_requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028898966/work
brotlipy==0.7.0
certifi==2022.5.18.1
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.13.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1649105784853/work
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
dateparser==1.1.1
frozenlist==1.3.0
fs==2.4.16
greenlet==1.1.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict==6.0.2
numexpr @ file:///tmp/build/80754af9/numexpr_1640689833592/work
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.2
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-binance==1.0.16
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==1.4.36
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
tzdata==2022.1
tzlocal==4.2
ujson==5.3.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
websockets==10.3
yarl==1.7.2
